# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.6.0
  description: Recall API — health, auth, notebooks, notes, tasks, capture, projects, and AI operations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecallTask"
  /tasks/bulk:
    post:
      operationId: bulkUpsertTasks
      tags: [tasks]
      summary: Bulk create or update tasks (migration)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkTasksRequest"
      responses:
        "200":
          description: Upserted tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecallTask"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Task not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RecallTask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        time:
          type: string
          nullable: true
        priority:
          type: string
          enum: [high, med, medium, low, none]
        tags:
          type: array
          items:
            type: string
        completed:
          type: boolean
        projectId:
          type: string
          nullable: true
      required:
        - id
        - title
        - priority
        - completed
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/RecallTask"
      required:
        - tasks
    CreateTaskRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        time:
          type: string
          nullable: true
        priority:
          type: string
          enum: [high, med, medium, low, none]
        tags:
          type: array
          items:
            type: string
        completed:
          type: boolean
        projectId:
          type: string
          nullable: true
      required:
        - title
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        time:
          type: string
          nullable: true
        priority:
          type: string
          enum: [high, med, medium, low, none]
        tags:
          type: array
          items:
            type: string
        completed:
          type: boolean
        projectId:
          type: string
          nullable: true
    BulkTasksRequest:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/CreateTaskRequest"
          maxItems: 500
      required:
        - tasks